+++ packages/global/core/workflow/runtime/type.ts
@@ -220,17 +220,8 @@ export const DispatchNodeResponseSchema = z
     limit: z.number().optional().meta({ description: '限制' }),
     searchMode: z.enum(DatasetSearchModeEnum).optional().meta({ description: '搜索模式' }),
     embeddingWeight: z.number().optional().meta({ description: '嵌入权重' }),
-    filteredFileCount: z.number().optional().meta({ description: '过滤的非图片文件数量' }),
-    queryImages: z
-      .array(
-        z.object({
-          key: z.string().optional(),
-          url: z.string().optional(),
-          name: z.string().optional()
-        })
-      )
-      .optional()
-      .meta({ description: '参与知识库检索的图片' }),
+    datasetQueries: z.array(z.string()).optional().meta({ description: '检索词' }),
+
     rerankModel: z.string().optional().meta({ description: '重排模型' }),
     rerankWeight: z.number().optional().meta({ description: '重排权重' }),
     reRankInputTokens: z.number().optional().meta({ description: '重排输入 token' }),
