+++ packages/service/core/workflow/dispatch/ai/toolcall/index.ts
@@ -1,4 +1,4 @@
-import { NodeInputKeyEnum, NodeOutputKeyEnum } from '@fastgpt/global/core/workflow/constants';
+import { NodeOutputKeyEnum } from '@fastgpt/global/core/workflow/constants';
 import { DispatchNodeResponseKeyEnum } from '@fastgpt/global/core/workflow/runtime/constants';
 import type { DispatchNodeResultType } from '@fastgpt/global/core/workflow/runtime/type';
 import { getLLMModel } from '../../../../ai/model';
@@ -18,7 +18,7 @@ type Response = DispatchNodeResultType<{
 
 export const dispatchRunTools = async (props: DispatchToolModuleProps): Promise<Response> => {
   const {
-    node: { nodeId, isEntry, inputs },
+    node: { nodeId, isEntry },
     runtimeNodes,
     runtimeEdges,
     histories,
@@ -33,14 +33,13 @@ export const dispatchRunTools = async (props: DispatchToolModuleProps): Promise<
       systemPrompt,
       userChatInput,
       history = 6,
-      fileUrlList,
+      fileUrlList: fileLinks,
       aiChatVision,
       aiChatReasoning,
       isResponseAnswerText = true,
       useAgentSandbox
     }
   } = props;
-  let fileLinks = fileUrlList;
 
   const useSandbox = !!useAgentSandbox && !!global.feConfigs?.show_agent_sandbox;
 
@@ -51,18 +50,12 @@ export const dispatchRunTools = async (props: DispatchToolModuleProps): Promise<
 
     props.params.aiChatVision = aiChatVision && toolModel.vision;
     props.params.aiChatReasoning = aiChatReasoning && toolModel.reasoning;
-    const fileUrlInput = inputs.find((item) => item.key === NodeInputKeyEnum.fileUrlList);
-    const fileLinks =
-      !fileUrlInput || !fileUrlInput.value || fileUrlInput.value.length === 0
-        ? undefined
-        : fileUrlList;
 
     const toolNodes = useToolNodeList({
       nodeId,
       runtimeNodes,
       runtimeEdges
     });
-    const fileUrls = userFiles.map((file) => file.url);
 
     // 交互恢复入口会由子工具继续接管，父 ToolCall 节点本轮不再作为入口节点。
     props.node.isEntry = false;
@@ -118,7 +111,6 @@ export const dispatchRunTools = async (props: DispatchToolModuleProps): Promise<
         toolNodes,
         toolModel,
         messages: adaptMessages,
-        fileUrls,
         childrenInteractiveParams:
           lastInteractive?.type === 'toolChildrenInteractive' ? lastInteractive.params : undefined
       });
