+++ packages/service/test/core/workflow/dispatch/ai/agent/sub/dataset.test.ts
@@ -16,7 +16,7 @@ const {
   formatModelChars2PointsMock: vi.fn()
 }));
 
-vi.mock('@fastgpt/service/core/dataset/search/controller', () => ({
+vi.mock('@fastgpt/service/core/dataset/search', () => ({
   defaultSearchDatasetData: defaultSearchDatasetDataMock
 }));
