+++ packages/web/i18n/zh-CN/workflow.json
@@ -14,6 +14,10 @@
   "Variable_name": "变量名",
   "add_new_input": "新增输入",
   "add_new_output": "新增输出",
+  "agent.prompt_skill_tip": "@选择 Skill 及工具，\"/\"选择变量",
+  "agent.select_skill": "选择 Skill",
+  "agent.select_tool": "选择工具",
+  "agent.tools": "工具",
   "append_application_reply_to_history_as_new_context": "将该应用回复内容拼接到历史记录中，作为新的上下文返回",
   "application_call": "应用调用",
   "assigned_reply": "指定回复",
@@ -21,6 +25,7 @@
   "auth_tmb_id_tip": "开启后，对外发布该应用时，还会根据用户是否有该知识库权限进行知识库过滤。\n若未开启，则直接按配置的知识库进行检索，不进行权限过滤。",
   "auto_align": "自动对齐",
   "can_not_loop": "该节点不支持循环嵌套",
+  "can_not_parallel": "该节点不支持在并行执行节点中使用",
   "choose_another_application_to_call": "选择一个其他应用进行调用",
   "classification_result": "分类结果",
   "click_to_change_reference": "点击切换输入模式",
@@ -41,9 +46,15 @@
   "confirm_delete_field_tip": "确认删除该字段？",
   "contains": "包含",
   "content_to_retrieve": "需要检索的内容",
-  "content_to_search": "需要检索的内容",
+  "content_to_search": "检索词或检索语句",
   "context_menu.add_comment": "添加注释",
   "create_link_error": "创建链接异常",
+  "current_index": "当前下标",
+  "current_index_desc": "数组模式下当前元素的 0-based 下标。",
+  "current_item": "当前元素",
+  "current_item_desc": "数组模式下当前迭代处理的元素。",
+  "current_iteration": "当前循环次数",
+  "current_iteration_desc": "条件循环模式下的 1-based 迭代次数。",
   "custom_feedback": "自定义反馈",
   "custom_input": "自定义变量",
   "dataset_quote_role": "角色",
@@ -95,6 +106,8 @@
   "intro_knowledge_base_search_merge": "可以将多个知识库搜索结果进行合并输出。使用 RRF 的合并方式进行最终排序输出。",
   "intro_laf_function_call": "可以调用Laf账号下的云函数。",
   "intro_loop": "输入一个数组，遍历数组并将每一个数组元素作为输入元素，执行工作流。",
+  "intro_loop_run": "执行重复性工作流，并由循环结果决定是否进行下一次操作。",
+  "intro_parallel_run": "输入一个数组，并行执行工作流中的每一个数组元素，最后将结果汇总输出。",
   "intro_plugin_input": "可以配置插件需要哪些输入，利用这些输入来运行插件",
   "intro_question_classification": "根据用户的历史记录和当前问题判断该次提问的类型。可以添加多组问题类型，下面是一个模板例子:\n类型1: 打招呼\n类型2: 关于商品“使用”问题\n类型3: 关于商品“购买”问题\n类型4: 其他问题",
   "intro_question_optimization": "使用问题优化功能，可以提高知识库连续对话时搜索的精度。使用该功能后，会先利用 AI 根据上下文构建一个或多个新的检索词，这些检索词更利于进行知识库搜索。该模块已内置在知识库搜索模块中，如果您仅进行一次知识库搜索，可直接使用知识库内置的补全功能。",
@@ -121,53 +134,28 @@
   "less_than_or_equal_to": "小于等于",
   "loop": "批量执行",
   "loop_body": "循环体",
+  "loop_custom_outputs": "自定义输出",
+  "loop_custom_outputs_tip": "声明对外暴露的输出字段，每个字段引用子流程内某个节点的输出。",
   "loop_end": "结束",
   "loop_input_array": "数组",
   "loop_result": "数组执行结果",
-  "loop_start": "开始",
-  "parallel_run": "并行执行",
-  "intro_parallel_run": "输入一个数组，并行执行工作流中的每一个数组元素，最后将结果汇总输出。",
-  "parallel_run_execution_logic": "执行逻辑",
-  "parallel_run_end_intro": "选择变量，作为并行执行的结果输出",
-  "parallel_run_max_concurrency": "最大并发数",
-  "parallel_run_max_concurrency_tip": "同时并行执行的最大任务数，范围 1～上限值（默认 5）。",
-  "parallel_run_max_retry_times": "单轮报错重试次数",
-  "parallel_run_max_retry_times_tip": "单个任务失败后的最大重试次数，范围 0～5（默认 3）。设为 0 表示失败不重试。",
-  "can_not_parallel": "该节点不支持在并行执行节点中使用",
-  "parallel_success_results": "成功结果",
-  "parallel_success_results_desc": "所有执行成功的任务输出，按输入顺序排列（仅包含成功项）。",
-  "parallel_full_results": "完整结果",
-  "parallel_full_results_desc": "与输入数组等长的结果数组，每项形如 {success, message, data}：成功时 success=true、message 为空、data 为输出值；失败时 success=false、message 为错误信息、data 为 null。",
-  "parallel_status": "完成状态",
-  "parallel_status_desc": "整体执行状态：success（全部成功）、partial_success（部分失败）、failed（全部失败）。",
-  "parallel_task": "任务 {{index}}",
-  "parallel_task_not_reach_end": "子工作流未到达结束节点",
-  "parallel_task_interactive_not_supported": "子工作流触发了交互式响应，并行执行节点不支持交互式节点",
   "loop_run": "循环节点",
-  "intro_loop_run": "执行重复性工作流，并由循环结果决定是否进行下一次操作。",
+  "loop_run_break": "循环终止",
+  "loop_run_break_must_inside_loop_run": "循环终止节点只能放在循环执行节点内部",
+  "loop_run_break_tip": "该模块需要在循环节点内部使用。当该模块被执行时，将强制结束循环节点的运行。",
+  "loop_run_conditional_requires_break": "条件循环必须至少包含一个循环终止节点",
+  "loop_run_input_array": "数组",
+  "loop_run_input_not_array": "循环输入值不是数组",
+  "loop_run_iteration_failed": "循环执行失败",
+  "loop_run_max_iterations_exceeded": "循环达最大上限",
   "loop_run_mode": "循环类型",
-  "loop_run_mode_tip": "选择循环驱动方式：按数组元素依次执行，或持续执行直到命中「循环终止」。",
   "loop_run_mode_array": "数组循环",
   "loop_run_mode_array_desc": "按顺序处理列表内每一项",
   "loop_run_mode_conditional": "条件循环",
   "loop_run_mode_conditional_desc": "重复运行直至满足特定条件",
-  "loop_run_input_array": "数组",
-  "loop_custom_outputs": "自定义输出",
-  "loop_custom_outputs_tip": "声明对外暴露的输出字段，每个字段引用子流程内某个节点的输出。",
-  "loop_run_break": "循环终止",
-  "loop_run_break_tip": "该模块需要在循环节点内部使用。当该模块被执行时，将强制结束循环节点的运行。",
-  "loop_run_break_must_inside_loop_run": "循环终止节点只能放在循环执行节点内部",
-  "loop_run_conditional_requires_break": "条件循环必须至少包含一个循环终止节点",
-  "loop_run_input_not_array": "循环输入值不是数组",
-  "loop_run_iteration_failed": "循环执行失败",
-  "loop_run_max_iterations_exceeded": "循环达最大上限",
+  "loop_run_mode_tip": "选择循环驱动方式：按数组元素依次执行，或持续执行直到命中「循环终止」。",
   "loop_run_start": "循环开始",
-  "current_index": "当前下标",
-  "current_index_desc": "数组模式下当前元素的 0-based 下标。",
-  "current_item": "当前元素",
-  "current_item_desc": "数组模式下当前迭代处理的元素。",
-  "current_iteration": "当前循环次数",
-  "current_iteration_desc": "条件循环模式下的 1-based 迭代次数。",
+  "loop_start": "开始",
   "max_dialog_rounds": "最多携带多少轮对话记录",
   "max_tokens": "最大 Tokens",
   "mouse_priority": "鼠标优先\n- 左键按下后可拖动画布\n- 按住 Shift 并拖动可框选多个节点",
@@ -179,6 +167,23 @@
   "optional_value_type": "可选的数据类型",
   "optional_value_type_tip": "可以指定 1 个或多个数据类型，用户在动态添加字段时，仅可选择配置的类型",
   "pan_priority": "触摸板优先\n- 单指拖拽框选节点\n- 双指滑动移动画布\n- 三指可快速移动节点",
+  "parallel_full_results": "完整结果",
+  "parallel_full_results_desc": "与输入数组等长的结果数组，每项形如 {success, message, data}：成功时 success=true、message 为空、data 为输出值；失败时 success=false、message 为错误信息、data 为 null。",
+  "parallel_run": "并行执行",
+  "parallel_run_end_intro": "选择变量，作为并行执行的结果输出",
+  "parallel_run_execution_logic": "执行逻辑",
+  "parallel_run_max_concurrency": "最大并发数",
+  "parallel_run_max_concurrency_tip": "同时并行执行的最大任务数，范围 1～上限值（默认 5）。",
+  "parallel_run_max_retry_times": "单轮报错重试次数",
+  "parallel_run_max_retry_times_tip": "单个任务失败后的最大重试次数，范围 0～5（默认 3）。设为 0 表示失败不重试。",
+  "parallel_status": "完成状态",
+  "parallel_status_desc": "整体执行状态：success（全部成功）、partial_success（部分失败）、failed（全部失败）。",
+  "parallel_success_results": "成功结果",
+  "parallel_success_results_desc": "所有执行成功的任务输出，按输入顺序排列（仅包含成功项）。",
+  "parallel_task": "任务 {{index}}",
+  "parallel_task_interactive_not_supported": "子工作流触发了交互式响应，并行执行节点不支持交互式节点",
+  "parallel_task_not_reach_end": "子工作流未到达结束节点",
+  "params_setting": "参数设置",
   "pass_returned_object_as_output_to_next_nodes": "将代码中 return 的对象作为输出，传递给后续的节点。变量名需要对应 return 的 key",
   "please_enter_node_name": "请输入节点名称",
   "plugin.Instruction_Tip": "可以配置一段说明，以解释该插件的用途。每次使用插件前，会显示该段说明。支持标准 Markdown 语法。",
@@ -205,6 +210,7 @@
   "response.Error": "错误信息",
   "response.Read file result": "文档解析结果预览",
   "response.read files": "解析的文档",
+  "search_query": "检索内容",
   "select_an_application": "选择一个应用",
   "select_another_application_to_call": "可以选择一个其他应用进行调用",
   "select_default_option": "选择默认值",
@@ -215,10 +221,6 @@
   "template.agent_intro": "由 AI 自主决定工具调用。",
   "template.agent_module": "Agent",
   "template.agent_module_intro": "关联所需工具及 Skill，可实现 AI 自主调用与规划",
-  "agent.tools": "工具",
-  "agent.prompt_skill_tip": "@选择 Skill 及工具，\"/\"选择变量",
-  "agent.select_skill": "选择 Skill",
-  "agent.select_tool": "选择工具",
   "template.ai_chat": "AI 对话",
   "template.ai_chat_intro": "AI 大模型对话",
   "template.dataset_search": "知识库搜索",
@@ -271,6 +273,5 @@
   "workflow.My edit": "我的编辑",
   "workflow.Switch_success": "切换成功",
   "workflow.Team cloud": "团队云端",
-  "workflow.exit_tips": "您的更改尚未保存，「直接退出」将不会保存您的编辑记录。",
-  "params_setting": "参数设置"
+  "workflow.exit_tips": "您的更改尚未保存，「直接退出」将不会保存您的编辑记录。"
 }
