+++ packages/web/i18n/zh-Hant/workflow.json
@@ -14,6 +14,10 @@
   "Variable_name": "變數名稱",
   "add_new_input": "新增輸入",
   "add_new_output": "新增輸出",
+  "agent.prompt_skill_tip": "@選擇 Skill 及工具，\"/\"選擇變數",
+  "agent.select_skill": "選擇 Skill",
+  "agent.select_tool": "選擇工具",
+  "agent.tools": "工具",
   "append_application_reply_to_history_as_new_context": "將應用程式的回覆附加到歷史紀錄中，作為新的脈絡",
   "application_call": "應用程式呼叫",
   "assigned_reply": "指定回覆",
@@ -21,6 +25,7 @@
   "auth_tmb_id_tip": "開啟後，對外發布應用程式時，也會根據使用者是否有該知識庫權限進行知識庫過濾。\n\n若未開啟，則直接按設定的知識庫進行檢索，不進行權限過濾。",
   "auto_align": "自動對齊",
   "can_not_loop": "這個節點不能迴圈。",
+  "can_not_parallel": "該節點不支援在並行執行節點中使用",
   "choose_another_application_to_call": "選擇另一個應用程式來呼叫",
   "classification_result": "分類結果",
   "click_to_change_reference": "點擊切換輸入模式",
@@ -41,9 +46,15 @@
   "confirm_delete_field_tip": "確定要刪除這個欄位嗎？",
   "contains": "包含",
   "content_to_retrieve": "要擷取的內容",
-  "content_to_search": "要搜尋的內容",
+  "content_to_search": "檢索字或檢索語句",
   "context_menu.add_comment": "新增註解",
   "create_link_error": "建立連結發生錯誤",
+  "current_index": "目前索引",
+  "current_index_desc": "陣列模式下目前元素的 0-based 索引。",
+  "current_item": "目前元素",
+  "current_item_desc": "陣列模式下目前迭代處理的元素。",
+  "current_iteration": "目前迴圈次數",
+  "current_iteration_desc": "條件迴圈模式下的 1-based 迭代次數。",
   "custom_feedback": "自訂回饋",
   "custom_input": "自定義變量",
   "dataset_quote_role": "角色",
@@ -95,6 +106,8 @@
   "intro_knowledge_base_search_merge": "可以合併多個知識庫搜尋結果並輸出。使用 RRF 合併方法進行最終排序輸出。",
   "intro_laf_function_call": "可以呼叫 Laf 帳號下的雲端函式。",
   "intro_loop": "輸入一個陣列，遍歷陣列並將每個陣列元素作為輸入元素，執行工作流程。",
+  "intro_loop_run": "執行重複性工作流，由迴圈結果決定是否進行下一次操作。",
+  "intro_parallel_run": "輸入一個陣列，並行執行工作流中的每一個陣列元素，最後將結果匯總輸出。",
   "intro_plugin_input": "可以設定外掛程式需要的輸入，並利用這些輸入來執行外掛程式",
   "intro_question_classification": "根據使用者的歷史紀錄和目前問題判斷這次提問的類型。可以新增多個問題類型，以下是一個範例：\n類型 1：打招呼\n類型 2：關於產品「使用方式」的問題\n類型 3：關於產品「購買」的問題\n類型 4：其他問題",
   "intro_question_optimization": "使用問題最佳化功能，可以提升知識庫連續對話時的搜尋精準度。使用這個功能後，會先利用 AI 根據脈絡建構一個或多個新的檢索詞彙，這些詞彙更有利於知識庫搜尋。這個模組已內建於知識庫搜尋模組中，如果您只進行一次知識庫搜尋，可以直接使用知識庫內建的自動完成功能。",
@@ -121,53 +134,28 @@
   "less_than_or_equal_to": "小於或等於",
   "loop": "大量執行",
   "loop_body": "迴圈主體",
+  "loop_custom_outputs": "自訂輸出",
+  "loop_custom_outputs_tip": "宣告對外曝露的輸出欄位，每個欄位引用子流程內某個節點的輸出。",
   "loop_end": "結束",
   "loop_input_array": "陣列",
   "loop_result": "陣列執行結果",
-  "loop_start": "開始",
-  "parallel_run": "並行執行",
-  "intro_parallel_run": "輸入一個陣列，並行執行工作流中的每一個陣列元素，最後將結果匯總輸出。",
-  "parallel_run_execution_logic": "執行邏輯",
-  "parallel_run_end_intro": "選擇變數，作為並行執行的結果輸出",
-  "parallel_run_max_concurrency": "最大並發數",
-  "parallel_run_max_concurrency_tip": "同時並行執行的最大任務數，範圍 1～上限值（預設 5）。",
-  "parallel_run_max_retry_times": "單輪報錯重試次數",
-  "parallel_run_max_retry_times_tip": "單個任務失敗後的最大重試次數，範圍 0～5（預設 3）。設為 0 表示失敗不重試。",
-  "can_not_parallel": "該節點不支援在並行執行節點中使用",
-  "parallel_success_results": "成功結果",
-  "parallel_success_results_desc": "所有執行成功的任務輸出，按輸入順序排列（僅包含成功項）。",
-  "parallel_full_results": "完整結果",
-  "parallel_full_results_desc": "與輸入陣列等長的結果陣列，每項形如 {success, message, data}：成功時 success=true、message 為空、data 為輸出值；失敗時 success=false、message 為錯誤訊息、data 為 null。",
-  "parallel_status": "完成狀態",
-  "parallel_status_desc": "整體執行狀態：success（全部成功）、partial_success（部分失敗）、failed（全部失敗）。",
-  "parallel_task": "任務 {{index}}",
-  "parallel_task_not_reach_end": "子工作流未到達結束節點",
-  "parallel_task_interactive_not_supported": "子工作流觸發了互動式回應，並行執行節點不支援互動式節點",
   "loop_run": "迴圈節點",
-  "intro_loop_run": "執行重複性工作流，由迴圈結果決定是否進行下一次操作。",
+  "loop_run_break": "迴圈終止",
+  "loop_run_break_must_inside_loop_run": "迴圈終止節點只能放在迴圈執行節點內部",
+  "loop_run_break_tip": "此模組需要在迴圈節點內部使用。當此模組被執行時，將強制結束迴圈節點的執行。",
+  "loop_run_conditional_requires_break": "條件迴圈必須至少包含一個迴圈終止節點",
+  "loop_run_input_array": "陣列",
+  "loop_run_input_not_array": "迴圈輸入值不是陣列",
+  "loop_run_iteration_failed": "迴圈執行失敗",
+  "loop_run_max_iterations_exceeded": "迴圈達最大上限",
   "loop_run_mode": "迴圈類型",
-  "loop_run_mode_tip": "選擇迴圈驅動方式：依陣列元素逐一執行，或持續執行直到命中「迴圈終止」。",
   "loop_run_mode_array": "陣列迴圈",
   "loop_run_mode_array_desc": "依序處理清單內每一項",
   "loop_run_mode_conditional": "條件迴圈",
   "loop_run_mode_conditional_desc": "重複執行直到滿足特定條件",
-  "loop_run_input_array": "陣列",
-  "loop_custom_outputs": "自訂輸出",
-  "loop_custom_outputs_tip": "宣告對外曝露的輸出欄位，每個欄位引用子流程內某個節點的輸出。",
-  "loop_run_break": "迴圈終止",
-  "loop_run_break_tip": "此模組需要在迴圈節點內部使用。當此模組被執行時，將強制結束迴圈節點的執行。",
-  "loop_run_break_must_inside_loop_run": "迴圈終止節點只能放在迴圈執行節點內部",
-  "loop_run_conditional_requires_break": "條件迴圈必須至少包含一個迴圈終止節點",
-  "loop_run_input_not_array": "迴圈輸入值不是陣列",
-  "loop_run_iteration_failed": "迴圈執行失敗",
-  "loop_run_max_iterations_exceeded": "迴圈達最大上限",
+  "loop_run_mode_tip": "選擇迴圈驅動方式：依陣列元素逐一執行，或持續執行直到命中「迴圈終止」。",
   "loop_run_start": "迴圈開始",
-  "current_index": "目前索引",
-  "current_index_desc": "陣列模式下目前元素的 0-based 索引。",
-  "current_item": "目前元素",
-  "current_item_desc": "陣列模式下目前迭代處理的元素。",
-  "current_iteration": "目前迴圈次數",
-  "current_iteration_desc": "條件迴圈模式下的 1-based 迭代次數。",
+  "loop_start": "開始",
   "max_dialog_rounds": "最多攜帶幾輪對話紀錄",
   "max_tokens": "最大 Token 數",
   "mouse_priority": "滑鼠優先\n- 按下左鍵拖曳畫布\n- 按住 Shift 並拖曳可框選多個節點",
@@ -179,6 +167,23 @@
   "optional_value_type": "可選的資料類型",
   "optional_value_type_tip": "可以指定一或多個資料類型，使用者在動態新增欄位時，只能選擇已設定的類型",
   "pan_priority": "觸摸板優先\n- 單指拖拽框選節點\n- 雙指滑動移動畫布\n- 三指可快速移動節點",
+  "parallel_full_results": "完整結果",
+  "parallel_full_results_desc": "與輸入陣列等長的結果陣列，每項形如 {success, message, data}：成功時 success=true、message 為空、data 為輸出值；失敗時 success=false、message 為錯誤訊息、data 為 null。",
+  "parallel_run": "並行執行",
+  "parallel_run_end_intro": "選擇變數，作為並行執行的結果輸出",
+  "parallel_run_execution_logic": "執行邏輯",
+  "parallel_run_max_concurrency": "最大並發數",
+  "parallel_run_max_concurrency_tip": "同時並行執行的最大任務數，範圍 1～上限值（預設 5）。",
+  "parallel_run_max_retry_times": "單輪報錯重試次數",
+  "parallel_run_max_retry_times_tip": "單個任務失敗後的最大重試次數，範圍 0～5（預設 3）。設為 0 表示失敗不重試。",
+  "parallel_status": "完成狀態",
+  "parallel_status_desc": "整體執行狀態：success（全部成功）、partial_success（部分失敗）、failed（全部失敗）。",
+  "parallel_success_results": "成功結果",
+  "parallel_success_results_desc": "所有執行成功的任務輸出，按輸入順序排列（僅包含成功項）。",
+  "parallel_task": "任務 {{index}}",
+  "parallel_task_interactive_not_supported": "子工作流觸發了互動式回應，並行執行節點不支援互動式節點",
+  "parallel_task_not_reach_end": "子工作流未到達結束節點",
+  "params_setting": "參數設定",
   "pass_returned_object_as_output_to_next_nodes": "將程式碼中 return 的物件作為輸出，傳遞給後續的節點。變數名稱需要對應 return 的鍵值",
   "please_enter_node_name": "請輸入節點名稱",
   "plugin.Instruction_Tip": "您可以設定一段說明來解釋這個外掛程式的用途。每次使用外掛程式前，都會顯示這段說明。支援標準 Markdown 語法。",
@@ -205,6 +210,7 @@
   "response.Error": "錯誤訊息",
   "response.Read file result": "檔案解析結果預覽",
   "response.read files": "解析的檔案",
+  "search_query": "檢索內容",
   "select_an_application": "選擇一個應用程式",
   "select_another_application_to_call": "可以選擇另一個應用程式來呼叫",
   "select_default_option": "選擇默認值",
@@ -215,10 +221,6 @@
   "template.agent_intro": "透過 AI 模型自動選擇一或多個功能區塊進行呼叫，也可以呼叫外掛程式。",
   "template.agent_module": "Agent",
   "template.agent_module_intro": "關聯所需工具及 Skill，可實現 AI 自主呼叫與規劃",
-  "agent.tools": "工具",
-  "agent.prompt_skill_tip": "@選擇 Skill 及工具，\"/\"選擇變數",
-  "agent.select_skill": "選擇 Skill",
-  "agent.select_tool": "選擇工具",
   "template.ai_chat": "AI 對話",
   "template.ai_chat_intro": "AI 大型語言模型對話",
   "template.dataset_search": "知識庫搜尋",
@@ -271,6 +273,5 @@
   "workflow.My edit": "我的編輯",
   "workflow.Switch_success": "切換成功",
   "workflow.Team cloud": "團隊雲端",
-  "workflow.exit_tips": "您的變更尚未儲存，「直接結束」將不會儲存您的編輯紀錄。",
-  "params_setting": "參數設定"
+  "workflow.exit_tips": "您的變更尚未儲存，「直接結束」將不會儲存您的編輯紀錄。"
 }
