+++ projects/app/src/components/core/chat/components/WholeResponseModal/ImageQuery.tsx
@@ -1,17 +1,43 @@
 import ImagePreviewToken from '@/components/core/dataset/ImagePreviewToken';
+import type { ImagePreviewTokenItemType } from '@/components/core/dataset/ImagePreviewToken';
 import { Box } from '@chakra-ui/react';
-import type { ChatHistoryItemResType } from '@fastgpt/global/core/chat/type';
 import React from 'react';
 
+const httpUrlReg = /^https?:\/\//i;
+
+/**
+ * nodeResponse 里为了兼容历史展示，把文本 query 和图片 URL 都放在 datasetQueries。
+ * UI 侧只负责展示拆分：普通 http(s) URL 作为图片 token，其余内容按原文本换行展示。
+ */
+const splitDatasetQueries = (datasetQueries: string[]) => {
+  const textQueries: string[] = [];
+  const queryImages: ImagePreviewTokenItemType[] = [];
+
+  for (const query of datasetQueries) {
+    if (!query) continue;
+
+    if (httpUrlReg.test(query)) {
+      queryImages.push({ url: query });
+    } else {
+      textQueries.push(query);
+    }
+  }
+
+  return {
+    query: textQueries.join('\n'),
+    queryImages
+  };
+};
+
 const ImageQuery = ({
-  query,
-  queryImages,
+  datasetQueries,
   datasetId
 }: {
-  query?: string;
-  queryImages: NonNullable<ChatHistoryItemResType['queryImages']>;
+  datasetQueries: string[];
   datasetId?: string;
 }) => {
+  const { query, queryImages } = splitDatasetQueries(datasetQueries);
+
   return (
     <Box
       border={'1px solid'}
